refactor(test-support): STOMP 세션 생성 로직을 TestStompSessionFactory로 추출
WebSocketIntegrationTestSupport·RoomModuleWebSocketTest·WebSocketConnectionTest에
중복된 SockJsClient/WebSocketStompClient 셋업 코드를 :test-support의
TestStompSessionFactory.connect(StompHeaders)로 통합
